# [aws-cpp-sdk-s3-crt] Control memory usage of downloads