@jmklix 2023-05-15
Thanks for making this PR. Merged with code-gen changes [here](https://github.com/aws/aws-sdk-cpp/pull/2491)